# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * eFoods - Food Ordering & Delivery Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MessageResponse {
  message: string;
}

export interface RegisterInput {
  /** @minLength 1 */
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
  phone?: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  customer: 'customer',
  restaurant_owner: 'restaurant_owner',
  admin: 'admin',
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  role: UserRole;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface Restaurant {
  id: number;
  name: string;
  description: string;
  cuisine: string;
  rating: number;
  reviewCount?: number;
  deliveryTime: string;
  deliveryFee: number;
  minOrder: number;
  imageUrl: string;
  isOpen: boolean;
  /** @nullable */
  address?: string | null;
  createdAt: string;
}

export interface RestaurantInput {
  /** @minLength 1 */
  name: string;
  description: string;
  cuisine: string;
  deliveryTime: string;
  deliveryFee: number;
  minOrder: number;
  imageUrl: string;
  address?: string;
}

export interface MenuItem {
  id: number;
  restaurantId: number;
  name: string;
  description: string;
  price: number;
  category: string;
  imageUrl: string;
  isAvailable: boolean;
  isVeg: boolean;
  isBestseller: boolean;
  /** @nullable */
  rating?: number | null;
  orderCount?: number;
}

export interface MenuItemInput {
  /** @minLength 1 */
  name: string;
  description: string;
  /** @minimum 0 */
  price: number;
  category: string;
  imageUrl: string;
  isVeg?: boolean;
  isBestseller?: boolean;
}

export interface MenuCategory {
  name: string;
  icon: string;
  itemCount: number;
}

export interface CartItem {
  id: number;
  menuItemId: number;
  name: string;
  price: number;
  quantity: number;
  imageUrl: string;
  isVeg?: boolean;
}

export interface Cart {
  items: CartItem[];
  subtotal: number;
  deliveryFee: number;
  total: number;
  /** @nullable */
  restaurantId: number | null;
  /** @nullable */
  restaurantName?: string | null;
}

export interface CartItemInput {
  menuItemId: number;
  /** @minimum 1 */
  quantity: number;
}

export interface CartItemUpdate {
  /** @minimum 0 */
  quantity: number;
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  preparing: 'preparing',
  out_for_delivery: 'out_for_delivery',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export interface OrderItem {
  id: number;
  menuItemId: number;
  name: string;
  price: number;
  quantity: number;
  imageUrl: string;
}

export interface Order {
  id: number;
  userId: number;
  restaurantId: number;
  restaurantName: string;
  items: OrderItem[];
  subtotal: number;
  deliveryFee: number;
  total: number;
  status: OrderStatus;
  deliveryAddress: string;
  /** @nullable */
  estimatedDelivery?: string | null;
  createdAt: string;
}

export interface OrderInput {
  deliveryAddress: string;
  razorpayOrderId: string;
  razorpayPaymentId: string;
  razorpaySignature: string;
}

export interface CategoryStat {
  category: string;
  count: number;
}

export interface StatusStat {
  status: string;
  count: number;
}

export interface DashboardSummary {
  totalOrders: number;
  totalRestaurants: number;
  totalUsers: number;
  popularCategories: CategoryStat[];
  ordersByStatus: StatusStat[];
}

export type ListRestaurantsParams = {
category?: string;
search?: string;
};