# lib/api-zod/src/generated/types/cartItem.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * eFoods - Food Ordering & Delivery Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface CartItem {
  id: number;
  menuItemId: number;
  name: string;
  price: number;
  quantity: number;
  imageUrl: string;
  isVeg?: boolean;
}